Fix bug with reordering scripts
When adding multiple scripts instead of 1 at a time, the index for each of them was the same. This has been fixed by calculating the index off of the local custom_scripts dict length instead of getting the length of the script tiles which are generated only after the for loop.